+++ sdk/communication/Azure.Communication.Identity/src/CommunicationIdentityClientOptions.cs
@@ -14,7 +14,7 @@ namespace Azure.Communication.Identity
         /// <summary>
         /// The latest version of the identity service.
         /// </summary>
-        internal const ServiceVersion LatestVersion = ServiceVersion.V2025_03_02_PREVIEW;
+        internal const ServiceVersion LatestVersion = ServiceVersion.V2026_09_23;
 
         internal string ApiVersion { get; }
 
@@ -29,7 +29,7 @@ namespace Azure.Communication.Identity
                 ServiceVersion.V2022_06_01 => "2022-06-01",
                 ServiceVersion.V2022_10_01 => "2022-10-01",
                 ServiceVersion.V2023_10_01 => "2023-10-01",
-                ServiceVersion.V2025_03_02_PREVIEW => "2025-03-02-preview",
+                ServiceVersion.V2026_09_23 => "2026-09-23",
                 _ => throw new ArgumentOutOfRangeException(nameof(version))
             };
         }
@@ -57,9 +57,9 @@ namespace Azure.Communication.Identity
             /// </summary>
             V2023_10_01 = 4,
             /// <summary>
-            /// The V2025_04_01_PREVIEW of the identity service.
+            /// The V2026_09_23 of the identity service.
             /// </summary>
-            V2025_03_02_PREVIEW = 5,
+            V2026_09_23 = 5,
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
     }